+++ ___tests___/testers/resolve/test/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-undef */
 const { get, getPath, find } = require('../../../../index');
+const resolve = require('../../../../src/handlers/get/resolve');
 
 /**
  * Check results get, vs find, vs getPath vs expected get result
@@ -7,10 +8,12 @@ const { get, getPath, find } = require('../../../../index');
  * @param {String/Array} path
  * @param {Array} results
  */
-const test = (input, path, result) => {
+// eslint-disable-next-line complexity
+const test = (input, path, result, expectedReferences) => {
   const getResult = get(input, path);
   const findResult = find(input, path);
   const pathResult = getPath(input, path);
+  const resolveResult = resolve(input, path);
   expect(getResult).toStrictEqual(result);
   if (!getResult) {
     expect(getResult).toStrictEqual(pathResult);
@@ -20,6 +23,10 @@ const test = (input, path, result) => {
   if (result !== undefined) {
     expect(get(input, pathResult)).toStrictEqual(result);
     expect(typeof pathResult).toStrictEqual('string');
+  }
+  expect(resolveResult.value).toStrictEqual(findResult.value);
+  if (expectedReferences) {
+    expect(resolveResult.references).toStrictEqual(expectedReferences);
   }
 };