+++ src/handlers/get/resolve.js
@@ -1,4 +1,5 @@
 const JsonGo = require('../../../index');
+
 const getSinglePathElement = require('../../helpers/pathElements/getSingle');
 const doesPathIndicateComplexity = require('./src/doesPathIndicateComplexity');
 const simpleGet = require('./src/simpleGet');
@@ -7,6 +8,8 @@ const getAllKeysFromObject = require('../../helpers/pathElements/getKeys/getAllK
 const getAllKeysFromArray = require('../../helpers/pathElements/getKeys/getAllKeysFromArray');
 const getTypeFromTempObject = require('./src/getTypeFromTempObject');
 const initializeGetHandler = require('./src/initializeGetHandler');
+const updateReferences = require('./src/updateReferences');
+const makeObject = require('../make/makeObject');
 
 /**
  * Handle wildcard, checks for each key whether remaining path is found and returns first key that
@@ -38,12 +41,14 @@ const getFirstKey = (
  * Get name of element either from wildcard or from other type of element
  */
 const getPathElement = (
-  element, obj, tempObject, getType, priorPath, arrayPath, index, functions, settings,
+  element, obj, tempObject, getType, priorPath, arrayPath, index, functions, settings, refObject,
 ) => {
   if (element.wildcard) {
     return getFirstKey(tempObject, arrayPath, getType, index, obj, priorPath, functions, settings);
   }
-  return getSinglePathElement(element, obj, tempObject, getType, priorPath, functions, settings);
+  return getSinglePathElement(
+    element, obj, tempObject, getType, priorPath, functions, settings, refObject,
+  );
 };
 
 /**
@@ -51,27 +56,29 @@ const getPathElement = (
  * iterations is desired
  */
 const getIteration = (
-  element, obj, tempObject, type, priorPath, arrayPath, index, functions, settingsObject,
+  element, obj, tempObject, type, priorPath, arrayPath, index, functions, settingsObject, refObject,
 ) => {
   let tempObj = tempObject;
   const arrPath = arrayPath;
   const elementValue = getPathElement(
-    element, obj, tempObject, type, priorPath, arrayPath, index, functions, settingsObject,
+    element, obj, tempObject, type, priorPath, arrayPath, index, functions,
+    settingsObject, refObject,
   );
   tempObj = tempObj[elementValue];
   arrPath[index] = elementValue;
   const priorPathElement = {};
   priorPathElement[type] = elementValue;
   priorPath.push(priorPathElement);
+  updateReferences(refObject, tempObj, element, elementValue);
   return validateOutput(tempObj, arrPath.length - 1 === index);
 };
 
 /**
  * returns value with corresponding path, if no value is found, no path is returned
  */
-const returnResult = (value, path) => {
+const returnResult = (value, path, references) => {
   if (value !== undefined) {
-    return { value, path };
+    return { value, path, references };
   }
   return { value: undefined, path: undefined };
 };
@@ -82,24 +89,27 @@ const returnResult = (value, path) => {
  * @param {any} path - string or array representation of path to set.
  * @param {Object} functions - object of functions that can be called within query.
  * @param {Object} settings - object with settings.
+ * @param {Object} referenceObject - object with (resolved) references.
  * @returns {any} returns value and resolved path found at specified path,
  * in case that multiple logical checks satisfy the first element will be returned
  */
-const resolve = (obj, path, functions, settings) => {
+const resolve = (obj, path, functions, settings, referenceObject) => {
+  const refObject = makeObject(referenceObject);
   const { settingsObject, arrayPath, priorPath } = initializeGetHandler(path, functions, settings);
   if (!doesPathIndicateComplexity(arrayPath)) {
-    return returnResult(simpleGet(obj, arrayPath), path);
+    return returnResult(simpleGet(obj, arrayPath), path, refObject);
   }
   let tempObject = obj;
   arrayPath.every((element, index) => {
     const type = getTypeFromTempObject(tempObject);
     const { shouldItContinue, newTempObject } = getIteration(
-      element, obj, tempObject, type, priorPath, arrayPath, index, functions, settingsObject,
+      element, obj, tempObject, type, priorPath, arrayPath, index, functions,
+      settingsObject, refObject,
     );
     tempObject = newTempObject;
     return shouldItContinue;
   });
-  return returnResult(tempObject, priorPath);
+  return returnResult(tempObject, priorPath, refObject);
 };
 
 module.exports = resolve;